Final Homework 19
Thanks for the feedback. #1 Tip doesn't work. I've tried a mac and a windows, and "hovering over desired element to initiate the popover" doesn't work. I can hover for hours, no popup happens. #2, I've been using Ranorex with varying degrees of success for weeks, but it didn't help me this time. I found a colleague to help me. He said the assert that the teachers used in class was a weird way to go about it, and gave me a different assert that actually works.